Add "org.spine3.samples" packages everywhere.
