--- registration/src/main/proto/lobby/registration/order/order.proto
@@ -1,44 +1,0 @@
-//
-// Copyright 2015, TeamDev Ltd. All rights reserved.
-//
-// Redistribution and use in source and/or binary forms, with or without
-// modification, must retain the above copyright notice and the following
-// disclaimer.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-// A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-// OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-// LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-// THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-syntax = "proto3";
-
-package lobby.registration.order;
-
-option java_package="sample.lobby.registration.order";
-option java_outer_classname = "OrderProto";
-option java_multiple_files = true;
-
-import "spine/samples/lobby/common/common.proto";
-import "spine/samples/lobby/registration/contracts/registration.proto";
-
-// The order made by a user (registrant) to reserve seats for a conference.
-message Order {
-    // The unique order id.
-    spine.samples.lobby.common.OrderId id = 1;
-
-    // The conference id the order associated with.
-    spine.samples.lobby.common.ConferenceId conference_id = 2;
-
-    // The set of seat items representing seat quantities of different types.
-    repeated spine.samples.lobby.registration.contracts.SeatQuantity seat = 3;
-
-    // The order is confirmed when the registrant has successfully paid for the order items.
-    bool is_confirmed = 4;
-}
