+++ registration/src/main/proto/spine/samples/lobby/registration/order/commands.proto
@@ -1,0 +1,81 @@
+//
+// Copyright 2015, TeamDev Ltd. All rights reserved.
+//
+// Redistribution and use in source and/or binary forms, with or without
+// modification, must retain the above copyright notice and the following
+// disclaimer.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+// A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+// OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+// LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+// THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
+syntax = "proto3";
+
+package spine.samples.lobby.registration.order;
+
+option java_package="org.spine3.samples.lobby.registration.order";
+option java_outer_classname = "OrderCommandsProto";
+option java_multiple_files = true;
+
+import "spine/samples/lobby/common/common.proto";
+import "spine/samples/lobby/registration/contracts/registration.proto";
+import "google/protobuf/timestamp.proto";
+
+// The request to reserve seats at the conference.
+// Is sent as the result of a registrant interaction with the UI.
+message RegisterToConference {
+    // The ID of the new order.
+    spine.samples.lobby.common.OrderId order_id = 1;
+
+    // The ID of the conference.
+    spine.samples.lobby.common.ConferenceId conference_id = 2;
+
+    // The order items representing the quantity of seats at the conference.
+    repeated spine.samples.lobby.registration.contracts.SeatQuantity seat = 3;
+}
+
+// The request to mark order items as reserved.
+// Is sent when a reservation is confirmed.
+message MarkSeatsAsReserved {
+    // The ID of the order.
+    spine.samples.lobby.common.OrderId order_id = 1;
+
+    // The time of the reservation expiration.
+    google.protobuf.Timestamp reservation_expiration = 2;
+
+    // The order items representing the quantity of seats of different types.
+    repeated spine.samples.lobby.registration.contracts.SeatQuantity seat = 3;
+}
+
+// The request to reject the order.
+// Is sent from a RegistrationProcessManager to an Order aggregate when the reservation is rejected or expired.
+message RejectOrder {
+    // The ID of the order to reject.
+    spine.samples.lobby.common.OrderId order_id = 1;
+}
+
+// The request to assign the registrant to the order.
+// Is sent as the result of a registrant interaction with the UI.
+message AssignRegistrantDetails {
+    // The ID of the order.
+    spine.samples.lobby.common.OrderId order_id = 1;
+
+    // The info about the registrant.
+    spine.samples.lobby.common.PersonalInfo registrant = 2;
+}
+
+// The request to confirm the order.
+// Is sent from a RegistrationProcessManager to an Order aggregate when the order is confirmed
+// (payment for the order is completed if the order is not free of charge).
+message ConfirmOrder {
+    // The ID of the order to confirm.
+    spine.samples.lobby.common.OrderId order_id = 1;
+}
