+++ registration/src/main/proto/spine/samples/lobby/registration/seat/assignment/seat_assignments.proto
@@ -1,0 +1,57 @@
+//
+// Copyright 2015, TeamDev Ltd. All rights reserved.
+//
+// Redistribution and use in source and/or binary forms, with or without
+// modification, must retain the above copyright notice and the following
+// disclaimer.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+// A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+// OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+// LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+// THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
+syntax = "proto3";
+
+package spine.samples.lobby.registration.seat.assignment;
+
+option java_package="org.spine3.samples.lobby.registration.seat.assignment";
+option java_outer_classname = "SeatAssignmentsProto";
+option java_multiple_files = true;
+
+import "spine/samples/lobby/common/common.proto";
+import "spine/samples/lobby/registration/contracts/registration.proto";
+
+// The state of the seat assignments aggregate.
+message SeatAssignments {
+    // The ID of the seat assignment.
+    SeatAssignmentsId id = 1;
+
+    // The map from a SeatPosition to a seat assignment.
+    // NOTE: the SeatPosition cannot be used as a map key due to Protobuf restrictions.
+    map<int32, SeatAssignment> seats = 2;
+}
+
+// The ID of the seat assignments aggregate.
+message SeatAssignmentsId {
+    // The UUID-based value of the seat assignments ID.
+    string uuid = 1;
+}
+
+// Associates an attendee with a seat in a confirmed order.
+message SeatAssignment {
+    // The position of the seat.
+    spine.samples.lobby.registration.contracts.SeatPosition position = 1;
+
+    // The seat type for the attendee.
+    spine.samples.lobby.common.SeatTypeId seat_type_id = 2;
+
+    // The personal info of the attendee.
+    spine.samples.lobby.common.PersonalInfo attendee = 3;
+}
