+++ registration/src/main/proto/spine/samples/lobby/registration/seat/availability/events.proto
@@ -1,0 +1,68 @@
+//
+// Copyright 2015, TeamDev Ltd. All rights reserved.
+//
+// Redistribution and use in source and/or binary forms, with or without
+// modification, must retain the above copyright notice and the following
+// disclaimer.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+// A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+// OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+// LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+// THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
+syntax = "proto3";
+
+package spine.samples.lobby.registration.seat.availability;
+
+option java_package="org.spine3.samples.lobby.registration.seat.availability";
+option java_outer_classname = "SeatsAvailabilityEventsProto";
+option java_multiple_files = true;
+
+import "spine/samples/lobby/common/common.proto";
+import "spine/samples/lobby/registration/contracts/registration.proto";
+
+// The event propagated when the availability of the seats is changed.
+message AvailableSeatsChanged  {
+    // The ID of the reservation.
+    spine.samples.lobby.common.ReservationId reservation_id = 1;
+
+    // The items representing the quantity of seats of different types.
+    repeated spine.samples.lobby.registration.contracts.SeatQuantity seat = 2;
+}
+
+// The event propagated when the reservation confirmation is received.
+message SeatsReserved {
+    // The ID of the conference.
+    spine.samples.lobby.common.ConferenceId conference_id = 1;
+
+    // The ID of the reservation.
+    spine.samples.lobby.common.ReservationId reservation_id = 2;
+
+    // The items representing the quantity of reserved seats.
+    repeated spine.samples.lobby.registration.contracts.SeatQuantity reservation_details = 3;
+
+    // The items representing the quantity of available seats.
+    repeated spine.samples.lobby.registration.contracts.SeatQuantity available_seats_changed = 4;
+}
+
+// The event propagated when the payment for the reservation is received and the reservation is committed.
+message SeatsReservationCommitted {
+    // The ID of the reservation.
+    spine.samples.lobby.common.ReservationId reservation_id = 1;
+}
+
+// The event propagated when the seats reservation is cancelled.
+message SeatsReservationCancelled {
+    // The ID of the reservation.
+    spine.samples.lobby.common.ReservationId reservation_id = 1;
+
+    // The items representing the quantity of available seats.
+    repeated spine.samples.lobby.registration.contracts.SeatQuantity available_seats_changed = 2;
+}
